// src/types/permissions.enum.ts
export enum permissionsVariant {
    delete = 'delete',//high
    edit = 'edit', //middle
    addMember = 'addMember',//base
    removeMember = 'removeMember',//high
    sendMessage = 'sendMessage', //base
    addMedia = 'addMedia', //base
    changeRole = 'changeRole',//high
    changeDiscussion = 'changeDiscussion',//middle
  }
  